services/user/requirements.txt
grpcio==1.78.1
grpcio-tools==1.78.1
psycopg==3.3.3
pydantic==2.12.5
pydantic-settings==2.13.1
bcrypt==4.2.1
